commit ajust

+++ src/main/java/com/agendaria/agendaria/domain/model/Customer.java
@@ -13,7 +13,7 @@ public class Customer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id; // Integer
 
     @Column(nullable = false)
     private String name;
