+++ src/main/java/com/agendaria/agendaria/repository/AppointmentRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
+public interface AppointmentRepository extends JpaRepository<Appointment, Integer> {
 
     List<Appointment> findByBusinessAndStartDateTimeBetween(
             Business business,
