+++ src/main/java/com/agendaria/agendaria/repository/BusinessRepository.java
@@ -3,5 +3,5 @@ package com.agendaria.agendaria.repository;
 import com.agendaria.agendaria.domain.model.Business;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusinessRepository extends JpaRepository<Business, Long> {
+public interface BusinessRepository extends JpaRepository<Business, Integer> {
 }
